backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5
